Inicio fluxo listing

+++ src/Backend/GameExchange.Infrastructe/Migrations/Versions/Version0000004.cs
@@ -12,7 +12,7 @@ namespace GameExchange.Infrastructe.Migrations.Versions
                 .WithColumn("Description").AsCustom("TEXT").NotNullable()
                 .WithColumn("Price").AsDecimal(8, 2).NotNullable()
                 .WithColumn("Status").AsInt32().NotNullable()
-                .WithColumn("GameId").AsInt32().NotNullable().ForeignKey("FK_Listings_Game_Id", "Games", "Id")
+                .WithColumn("GameId").AsInt64().NotNullable().ForeignKey("FK_Listings_Game_Id", "Games", "Id")
                 .WithColumn("SellerId").AsInt64().NotNullable().ForeignKey("FK_Listings_User_Id", "Users", "Id");
